requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.5b2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2.3
djangorestframework==3.12.4
Flask==1.0.2
gunicorn==20.1.0
idna==2.10
itsdangerous==0.24
Jinja2==2.11.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.8.1
peewee==3.3.4
psycopg2-binary==2.8.6
python-dateutil==1.5
pytz==2021.1
regex==2021.4.4
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.5
Werkzeug==0.15.5
whitenoise==5.2.0
